+++ packages/build-tools/src/steps/easFunctions.ts
@@ -26,10 +26,6 @@ import { createLaunchApplicationFunction } from './functions/launchApplication';
 import { createPrebuildBuildFunction } from './functions/prebuild';
 import { createReadAppConfigBuildFunction } from './functions/readAppConfig';
 import { createReadIpaInfoBuildFunction } from './functions/readIpaInfo';
-import {
-  createCleanupSubmissionFunction,
-  createPrepareSubmissionFunction,
-} from './functions/prepareSubmission';
 import { createReadPackageJsonBuildFunction } from './functions/readPackageJson';
 import { createRepackBuildFunction } from './functions/repack';
 import { createReportMaestroTestResultsFunction } from './functions/reportMaestroTestResults';
@@ -81,8 +77,6 @@ export function getEasFunctions(ctx: CustomBuildContext): BuildFunction[] {
     createInstallNodeModulesBuildFunction(),
     createPrebuildBuildFunction(),
     createReadIpaInfoBuildFunction(),
-    createPrepareSubmissionFunction(ctx),
-    createCleanupSubmissionFunction(),
     createDownloadBuildFunction(ctx),
     createInstallBuildFunction(),
     createLaunchApplicationFunction(),